+++ internal/observability/middleware_test.go
@@ -279,6 +279,7 @@ func TestMiddlewareLoggingWithApplicationErrorConfiguration(t *testing.T) {
 		ShardKey:        "shard01",
 		RoutingKey:      "routing-key",
 		RoutingDelegate: "routing-delegate",
+		CallerProcedure: "caller-procedure",
 		Body:            strings.NewReader("body"),
 	}
 
@@ -508,6 +509,7 @@ func TestMiddlewareLoggingWithApplicationErrorConfiguration(t *testing.T) {
 			logContext := append(
 				baseFields(),
 				zap.String("direction", string(_directionInbound)),
+				zap.String("sourceProcedure", req.CallerProcedure),
 				zap.String("rpcType", "Unary"),
 			)
 			logContext = append(logContext, tt.wantFields...)
@@ -529,6 +531,7 @@ func TestMiddlewareLoggingWithApplicationErrorConfiguration(t *testing.T) {
 			logContext := append(
 				baseFields(),
 				zap.String("direction", string(_directionOutbound)),
+				zap.String("sourceProcedure", req.CallerProcedure),
 				zap.String("rpcType", "Unary"),
 			)
 			logContext = append(logContext, tt.wantFields...)
@@ -553,6 +556,7 @@ func TestMiddlewareLoggingWithApplicationErrorConfiguration(t *testing.T) {
 			logContext := append(
 				baseFields(),
 				zap.String("direction", string(_directionInbound)),
+				zap.String("sourceProcedure", req.CallerProcedure),
 				zap.String("rpcType", "Oneway"),
 			)
 			logContext = append(logContext, tt.wantFields...)
@@ -571,6 +575,7 @@ func TestMiddlewareLoggingWithApplicationErrorConfiguration(t *testing.T) {
 			logContext := append(
 				baseFields(),
 				zap.String("direction", string(_directionOutbound)),
+				zap.String("sourceProcedure", req.CallerProcedure),
 				zap.String("rpcType", "Oneway"),
 			)
 			logContext = append(logContext, tt.wantFields...)
@@ -585,6 +590,86 @@ func TestMiddlewareLoggingWithApplicationErrorConfiguration(t *testing.T) {
 				Context: logContext,
 			}
 			assert.Equal(t, expected, getLog(t), "Unexpected log entry written.")
+		})
+	}
+}
+
+func TestMiddlewareLoggingDifferentCallerProcedureWithSameEdge(t *testing.T) {
+	defer stubTime()()
+
+	infoLevel := zapcore.InfoLevel
+	warnLevel := zapcore.WarnLevel
+
+	core, logs := observer.New(zapcore.DebugLevel)
+	mw := NewMiddleware(Config{
+		Logger:           zap.New(core),
+		Scope:            metrics.New().Scope(),
+		ContextExtractor: NewNopContextExtractor(),
+		Levels: LevelsConfig{
+			Default: DirectionalLevelsConfig{
+				Success:          &infoLevel,
+				ApplicationError: &warnLevel,
+				// Leave failure level as the default.
+			},
+		},
+	})
+
+	getLog := func(t *testing.T) observer.LoggedEntry {
+		entries := logs.TakeAll()
+		e := entries[len(entries)-1]
+		e.Entry.Time = time.Time{}
+		return e
+	}
+
+	for _, tt := range []struct {
+		desc string
+		req  transport.Request
+	}{
+		{
+			desc: "caller1",
+			req: transport.Request{
+				Caller:          "caller",
+				Service:         "service",
+				Encoding:        "raw",
+				Procedure:       "procedure",
+				CallerProcedure: "callerProcedure_1",
+			},
+		},
+	} {
+		logContext := []zapcore.Field{
+			zap.String("source", tt.req.Caller),
+			zap.String("dest", tt.req.Service),
+			zap.String("transport", unknownIfEmpty(tt.req.Transport)),
+			zap.String("procedure", tt.req.Procedure),
+			zap.String("encoding", string(tt.req.Encoding)),
+			zap.String("routingKey", tt.req.RoutingKey),
+			zap.String("routingDelegate", tt.req.RoutingDelegate),
+			zap.String("direction", string(_directionInbound)),
+			zap.String("sourceProcedure", tt.req.CallerProcedure),
+			zap.String("rpcType", "Unary"),
+			zap.Duration("latency", 0),
+			zap.Bool("successful", true),
+			zap.Skip(), // ContextExtractor
+		}
+
+		t.Run(tt.desc, func(t *testing.T) {
+			err := mw.Handle(
+				context.Background(),
+				&tt.req,
+				&transporttest.FakeResponseWriter{},
+				fakeHandler{},
+			)
+			assert.NoError(t, err, "Unexpected error from middleware.")
+
+			expected := observer.LoggedEntry{
+				Entry: zapcore.Entry{
+					Level:   zapcore.InfoLevel,
+					Message: "Handled inbound request.",
+				},
+				Context: logContext,
+			}
+			l := getLog(t)
+			assert.Equal(t, expected, l, "Unexpected log entry written.")
 		})
 	}
 }
@@ -867,6 +952,7 @@ func TestMiddlewareLoggingWithServerErrorConfiguration(t *testing.T) {
 			logContext := append(
 				baseFields(),
 				zap.String("direction", string(_directionInbound)),
+				zap.String("sourceProcedure", req.CallerProcedure),
 				zap.String("rpcType", "Unary"),
 			)
 			logContext = append(logContext, tt.wantFields...)
@@ -888,6 +974,7 @@ func TestMiddlewareLoggingWithServerErrorConfiguration(t *testing.T) {
 			logContext := append(
 				baseFields(),
 				zap.String("direction", string(_directionOutbound)),
+				zap.String("sourceProcedure", req.CallerProcedure),
 				zap.String("rpcType", "Unary"),
 			)
 			logContext = append(logContext, tt.wantFields...)
@@ -912,6 +999,7 @@ func TestMiddlewareLoggingWithServerErrorConfiguration(t *testing.T) {
 			logContext := append(
 				baseFields(),
 				zap.String("direction", string(_directionInbound)),
+				zap.String("sourceProcedure", req.CallerProcedure),
 				zap.String("rpcType", "Oneway"),
 			)
 			logContext = append(logContext, tt.wantFields...)
@@ -930,6 +1018,7 @@ func TestMiddlewareLoggingWithServerErrorConfiguration(t *testing.T) {
 			logContext := append(
 				baseFields(),
 				zap.String("direction", string(_directionOutbound)),
+				zap.String("sourceProcedure", req.CallerProcedure),
 				zap.String("rpcType", "Oneway"),
 			)
 			logContext = append(logContext, tt.wantFields...)
@@ -961,6 +1050,7 @@ func TestMiddlewareStreamingSuccess(t *testing.T) {
 			ShardKey:        "shard-key",
 			RoutingKey:      "routing-key",
 			RoutingDelegate: "routing-delegate",
+			CallerProcedure: "caller-procedure",
 		},
 	}
 
@@ -1025,6 +1115,7 @@ func TestMiddlewareStreamingSuccess(t *testing.T) {
 		logFields := func() []zapcore.Field {
 			return newZapFields(
 				zap.String("direction", string(_directionInbound)),
+				zap.String("sourceProcedure", req.Meta.CallerProcedure),
 				zap.String("rpcType", "Streaming"),
 				zap.Bool("successful", true),
 				zap.Skip(), // context extractor
@@ -1083,6 +1174,7 @@ func TestMiddlewareStreamingSuccess(t *testing.T) {
 		fields := func() []zapcore.Field {
 			return newZapFields(
 				zap.String("direction", string(_directionOutbound)),
+				zap.String("sourceProcedure", req.Meta.CallerProcedure),
 				zap.String("rpcType", "Streaming"),
 				zap.Bool("successful", true),
 				zap.Skip(), // context extractor
@@ -1219,6 +1311,7 @@ func TestMiddlewareStreamingLoggingErrorWithFailureConfiguration(t *testing.T) {
 
 				fields := newZapFields(
 					zap.String("direction", string(_directionInbound)),
+					zap.String("sourceProcedure", req.Meta.CallerProcedure),
 					zap.String("rpcType", "Streaming"),
 					zap.Bool("successful", false),
 					zap.Skip(), // context extractor
@@ -1271,6 +1364,7 @@ func TestMiddlewareStreamingLoggingErrorWithFailureConfiguration(t *testing.T) {
 		fields := func() []zapcore.Field {
 			return newZapFields(
 				zap.String("direction", string(_directionInbound)),
+				zap.String("sourceProcedure", req.Meta.CallerProcedure),
 				zap.String("rpcType", "Streaming"),
 				zap.Bool("successful", false),
 				zap.Skip(), // context extractor
@@ -1313,6 +1407,7 @@ func TestMiddlewareStreamingLoggingErrorWithFailureConfiguration(t *testing.T) {
 		fields := func() []zapcore.Field {
 			return newZapFields(
 				zap.String("direction", string(_directionOutbound)),
+				zap.String("sourceProcedure", req.Meta.CallerProcedure),
 				zap.String("rpcType", "Streaming"),
 				zap.Bool("successful", false),
 				zap.Skip(), // context extractor
@@ -1359,6 +1454,7 @@ func TestMiddlewareStreamingLoggingErrorWithFailureConfiguration(t *testing.T) {
 		fields := func() []zapcore.Field {
 			return newZapFields(
 				zap.String("direction", string(_directionOutbound)),
+				zap.String("sourceProcedure", req.Meta.CallerProcedure),
 				zap.String("rpcType", "Streaming"),
 				zap.Bool("successful", false),
 				zap.Skip(), // context extractor
@@ -1448,6 +1544,7 @@ func TestMiddlewareStreamingLoggingErrorWithFailureConfiguration(t *testing.T) {
 		logFields := func(err error) []zapcore.Field {
 			return newZapFields(
 				zap.String("direction", string(_directionInbound)),
+				zap.String("sourceProcedure", req.Meta.CallerProcedure),
 				zap.String("rpcType", "Streaming"),
 				zap.Bool("successful", true),
 				zap.Skip(), // context extractor
@@ -1516,6 +1613,7 @@ func TestMiddlewareStreamingLoggingErrorWithServerClientConfiguration(t *testing
 			ShardKey:        "shard-key",
 			RoutingKey:      "routing-key",
 			RoutingDelegate: "routing-delegate",
+			CallerProcedure: "caller-procedure",
 		},
 	}
 
@@ -1596,6 +1694,7 @@ func TestMiddlewareStreamingLoggingErrorWithServerClientConfiguration(t *testing
 
 				fields := newZapFields(
 					zap.String("direction", string(_directionInbound)),
+					zap.String("sourceProcedure", req.Meta.CallerProcedure),
 					zap.String("rpcType", "Streaming"),
 					zap.Bool("successful", false),
 					zap.Skip(), // context extractor
@@ -1875,6 +1974,7 @@ func TestUnaryInboundApplicationErrors(t *testing.T) {
 		zap.String("routingKey", req.RoutingKey),
 		zap.String("routingDelegate", req.RoutingDelegate),
 		zap.String("direction", string(_directionInbound)),
+		zap.String("sourceProcedure", req.CallerProcedure),
 		zap.String("rpcType", "Unary"),
 		zap.Duration("latency", 0),
 		zap.Bool("successful", false),